@@ -1,67 +1,50 @@
 {
-  "name": "moniast-frontend",
+  "name": "front",
   "version": "0.1.0",
-  "description": "",
-  "main": "src/app.js",
-  "scripts": {
-    "test": "jest",
-    "dev": "webpack-dev-server --hot --inline --host 0.0.0.0 --port 8090 --public --config webpack.dev.js",
-    "build": "webpack --config webpack.prod.js --progress -p",
-    "snyk-protect": "snyk protect",
-    "prepare": "npm run snyk-protect"
-  },
-  "keywords": [
-    "asterisk",
-    "monitor",
-    "ui"
-  ],
-  "author": "Kol007",
-  "license": "MIT",
+  "private": true,
   "dependencies": {
-    "axios": "^0.17.1",
-    "immutable": "^3.8.2",
+    "@fortawesome/fontawesome-svg-core": "^1.2.12",
+    "@fortawesome/free-solid-svg-icons": "^5.6.3",
+    "@fortawesome/react-fontawesome": "^0.1.4",
+    "axios": "^0.18.0",
+    "bootstrap": "^4.2.1",
+    "immutable": "^4.0.0-rc.12",
     "moment": "^2.23.0",
-    "prop-types": "^15.6.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-fontawesome": "^1.6.1",
-    "react-redux": "^5.0.6",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "reactstrap": "^5.0.0-alpha.3",
+    "prop-types": "^15.6.2",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
+    "react-redux": "^6.0.0",
+    "react-router": "^4.4.0-beta.7",
+    "react-router-dom": "^4.4.0-beta.7",
+    "react-scripts": "2.1.8",
+    "react-toastify": "^4.5.2",
+    "reactstrap": "^7.1.0",
     "recompose": "^0.30.0",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0",
-    "reselect": "^3.0.1",
-    "socket.io-client": "^2.0.4"
+    "redux": "^4.0.1",
+    "redux-immutable": "^4.0.0",
+    "redux-thunk": "^2.3.0",
+    "reselect": "^4.0.0",
+    "socket.io-client": "^2.2.0"
   },
+  "scripts": {
+    "start": "react-scripts start",
+    "startl": "NODE_PATH=./src react-scripts start",
+    "build": "react-scripts build",
+    "buildl": "NODE_PATH=./src react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ],
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^21.2.0",
-    "babel-loader": "^7.1.2",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "clean-webpack-plugin": "^0.1.17",
-    "css-loader": "^0.28.7",
-    "enzyme": "^3.4.1",
-    "file-loader": "^1.1.5",
-    "jest": "^21.2.1",
-    "path": "^0.12.7",
-    "react-addons-test-utils": "^15.6.2",
-    "react-test-renderer": "^16.0.0",
     "redux-logger": "^3.0.6",
-    "redux-mock-store": "^1.3.0",
-    "style-loader": "^0.19.0",
-    "uglifyjs-webpack-plugin": "^1.2.7",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.10.0",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-dev-server": "^3.1.14",
-    "webpack-merge": "^4.1.1",
-    "why-did-you-update": "^0.1.1"
+    "why-did-you-update": "^1.0.6"
   }
 }
