@@ -1,23 +1,23 @@
-const AMI           = require('../helpers/ami');
+const AMI = require('../helpers/ami');
 const getListEvents = require('./common').getListEvents;
 
-const SIP_STATUS                  = require('../helpers/constants').SIP_STATUS;
-const PJSIP_STATUS                = require('../helpers/constants').PJSIP_STATUS;
-const SIP_STATUS_IDLE             = require('../helpers/constants').SIP_STATUS_IDLE;
-const SIP_STATUS_IN_USE           = require('../helpers/constants').SIP_STATUS_IN_USE;
-const SIP_STATUS_BUSY             = require('../helpers/constants').SIP_STATUS_BUSY;
-const SIP_STATUS_UNAVAILABLE      = require('../helpers/constants').SIP_STATUS_UNAVAILABLE;
-const SIP_STATUS_RINGING          = require('../helpers/constants').SIP_STATUS_RINGING;
-const SIP_STATUS_ON_HOLD          = require('../helpers/constants').SIP_STATUS_ON_HOLD;
-const SIP_STATUS_CODE_IN_USE      = require('../helpers/constants').SIP_STATUS_CODE_IN_USE;
-const SIP_STATUS_CODE_BUSY        = require('../helpers/constants').SIP_STATUS_CODE_BUSY;
+const SIP_STATUS = require('../helpers/constants').SIP_STATUS;
+const PJSIP_STATUS = require('../helpers/constants').PJSIP_STATUS;
+const SIP_STATUS_IDLE = require('../helpers/constants').SIP_STATUS_IDLE;
+const SIP_STATUS_IN_USE = require('../helpers/constants').SIP_STATUS_IN_USE;
+const SIP_STATUS_BUSY = require('../helpers/constants').SIP_STATUS_BUSY;
+const SIP_STATUS_UNAVAILABLE = require('../helpers/constants').SIP_STATUS_UNAVAILABLE;
+const SIP_STATUS_RINGING = require('../helpers/constants').SIP_STATUS_RINGING;
+const SIP_STATUS_ON_HOLD = require('../helpers/constants').SIP_STATUS_ON_HOLD;
+const SIP_STATUS_CODE_IN_USE = require('../helpers/constants').SIP_STATUS_CODE_IN_USE;
+const SIP_STATUS_CODE_BUSY = require('../helpers/constants').SIP_STATUS_CODE_BUSY;
 const SIP_STATUS_CODE_UNAVAILABLE = require('../helpers/constants').SIP_STATUS_CODE_UNAVAILABLE;
-const SIP_STATUS_CODE_RINGING     = require('../helpers/constants').SIP_STATUS_CODE_RINGING;
-const SIP_STATUS_CODE_ON_HOLD     = require('../helpers/constants').SIP_STATUS_CODE_ON_HOLD;
+const SIP_STATUS_CODE_RINGING = require('../helpers/constants').SIP_STATUS_CODE_RINGING;
+const SIP_STATUS_CODE_ON_HOLD = require('../helpers/constants').SIP_STATUS_CODE_ON_HOLD;
 
-exports.getSip = async function (req, res, next) {
+exports.getSip = async function(req, res, next) {
   let pjSipPromise = getAllPjSipWithInfo();
-  let sipPromise   = getAllSipWithInfo();
+  let sipPromise = getAllSipWithInfo();
 
   Promise.all([sipPromise, pjSipPromise]).then(([sip, pjsip]) => {
     req.data = [...sip, ...pjsip];
@@ -25,13 +25,13 @@ exports.getSip = async function (req, res, next) {
   });
 };
 
-exports.getSipResponse = function (req, res, next) {
+exports.getSipResponse = function(req, res, next) {
   res.json(req.data);
 };
 
 function getAllSipWithInfo() {
   return new Promise((resolve, reject) => {
-    AMI.action({ action: 'SIPpeers' }, async function (err, res) {
+    AMI.action({ action: 'SIPpeers' }, async function(err, res) {
       if (!err) {
         let response = await getListEvents(AMI, res.actionid, 'PeerlistComplete');
 
@@ -42,11 +42,11 @@ function getAllSipWithInfo() {
           return new Promise((resolve, reject) => {
             AMI.action({ action: 'SIPshowpeer', Peer: item.objectname }, (err, res) => {
               let login = res.callerid ? res.callerid.match(/"(\w+)"/) : '';
-              login     = login ? login[1] : '';
+              login = login ? login[1] : '';
 
               resolve({
-                id     : item.objectname,
-                sip    : item.objectname,
+                id: item.objectname,
+                sip: item.objectname,
                 isTrunk: !login,
                 login,
                 status,
@@ -71,7 +71,7 @@ function getPjsipAdditionalInfo(item) {
   let promise;
 
   promise = new Promise((resolve, reject) => {
-    AMI.action({ action: 'PJSIPShowEndpoint', endpoint: item.objectname }, function (err, res) {
+    AMI.action({ action: 'PJSIPShowEndpoint', endpoint: item.objectname }, function(err, res) {
       let listener = AMI.on('managerevent', f);
 
       function f(evt) {
@@ -82,14 +82,14 @@ function getPjsipAdditionalInfo(item) {
           } else {
             if (evt.event === 'EndpointDetail') {
               let login = evt.callerid ? evt.callerid.match(/"(\w+)"/) : '';
-              login     = login ? login[1] : '';
+              login = login ? login[1] : '';
 
               const online = item.devicestate !== 'Unavailable';
               const status = PJSIP_STATUS[item.devicestate] || SIP_STATUS_UNAVAILABLE;
 
               result = {
-                id     : item.objectname,
-                sip    : item.objectname,
+                id: item.objectname,
+                sip: item.objectname,
                 isTrunk: !login,
                 login,
                 status,
@@ -107,7 +107,7 @@ function getPjsipAdditionalInfo(item) {
 
 function getAllPjSipWithInfo() {
   return new Promise((resolve, reject) => {
-    AMI.action({ action: 'PJSIPShowEndpoints' }, async function (err, res) {
+    AMI.action({ action: 'PJSIPShowEndpoints' }, async function(err, res) {
       if (!err) {
         let response = await getListEvents(AMI, res.actionid, 'EndpointListComplete');
         let promises = response.map(item => getPjsipAdditionalInfo(item));
