@@ -1,5 +1,8 @@
-const ROLE_CLIENT = 'Client';
-const ROLE_ADMIN = 'Admin';
+const ROLE_STRING_CLIENT = 'Client';
+const ROLE_STRING_ADMIN = 'Admin';
+
+const ROLE_CLIENT = 5;
+const ROLE_ADMIN = 10;
 
 const SIP_STATUS_CODE_IDLE = '0';
 const SIP_STATUS_CODE_IN_USE = '1';
@@ -16,13 +19,13 @@ const SIP_STATUS_RINGING = 'SIP_STATUS_RINGING';
 const SIP_STATUS_ON_HOLD = 'SIP_STATUS_ON_HOLD';
 
 const PJSIP_STATUS = {
-  'Unavailable': SIP_STATUS_UNAVAILABLE,
-  'Ringing': SIP_STATUS_RINGING,
-  'Onhold': SIP_STATUS_ON_HOLD,
-  'Inuse': SIP_STATUS_IN_USE,
-  'Not_inuse': SIP_STATUS_IN_USE,
-  'Ringinuse': SIP_STATUS_RINGING,
-  'Busy': SIP_STATUS_BUSY,
+  Unavailable: SIP_STATUS_UNAVAILABLE,
+  Ringing: SIP_STATUS_RINGING,
+  Onhold: SIP_STATUS_ON_HOLD,
+  Inuse: SIP_STATUS_IN_USE,
+  Not_inuse: SIP_STATUS_IN_USE,
+  Ringinuse: SIP_STATUS_RINGING,
+  Busy: SIP_STATUS_BUSY
 };
 
 const SIP_STATUS = {
@@ -34,7 +37,6 @@ const SIP_STATUS = {
   16: 'SIP_STATUS_ON_HOLD'
 };
 
-
 const QUEUE_MEMBER_STATUS = {
   '0': 'SIP_STATUS_UNAVAILABLE', //'AST_DEVICE_UNKNOWN',
   '1': 'SIP_STATUS_IDLE',
@@ -44,7 +46,7 @@ const QUEUE_MEMBER_STATUS = {
   '5': 'SIP_STATUS_UNAVAILABLE',
   '6': 'SIP_STATUS_RINGING',
   '7': 'SIP_STATUS_IN_USE', //'AST_DEVICE_RINGINUSE',
-  '8': 'SIP_STATUS_ON_HOLD',
+  '8': 'SIP_STATUS_ON_HOLD'
 };
 
 const CHANNEL_STATUS_DOWN_AVAILABLE = 'CHANNEL_STATUS_DOWN_AVAILABLE';
@@ -77,6 +79,8 @@ const CHANNEL_STATUS = {
 };
 
 module.exports = {
+  ROLE_STRING_CLIENT,
+  ROLE_STRING_ADMIN,
   ROLE_CLIENT,
   ROLE_ADMIN,
 
