@@ -1,32 +1,32 @@
 const AuthenticationController = require('../controllers/authentication');
-const UserController           = require('../controllers/user');
-const SipController            = require('../controllers/sip');
-const ChannelsController       = require('../controllers/channels');
-const QueueController          = require('../controllers/queue');
-const CommonController         = require('../controllers/common');
-const express                  = require('express');
-const passport                 = require('passport');
-const passportService          = require('../config/passport');
+const UserController = require('../controllers/user');
+const SipController = require('../controllers/sip');
+const ChannelsController = require('../controllers/channels');
+const QueueController = require('../controllers/queue');
+const CommonController = require('../controllers/common');
+const express = require('express');
+const passport = require('passport');
+const passportService = require('../config/passport');
 
 const ROLE_CLIENT = require('./constants').ROLE_CLIENT;
-const ROLE_ADMIN  = require('./constants').ROLE_ADMIN;
+const ROLE_ADMIN = require('./constants').ROLE_ADMIN;
 
 // Middleware to require login/auth
-const requireAuth  = passport.authenticate('jwt', { session: false });
+const requireAuth = passport.authenticate('jwt', { session: false });
 const requireLogin = passport.authenticate('local', { session: false });
 
 // Constants for role types
-const REQUIRE_ADMIN  = 'Admin',
-      REQUIRE_CLIENT = 'Client';
+const REQUIRE_ADMIN = 'Admin',
+  REQUIRE_CLIENT = 'Client';
 
-module.exports = function (app) {
+module.exports = function(app) {
   // Initializing route groups
-  const apiRoutes      = express.Router(),
-        sipRoutes      = express.Router(),
-        queueRoutes    = express.Router(),
-        channelsRoutes = express.Router(),
-        userRoutes     = express.Router(),
-        authRoutes     = express.Router();
+  const apiRoutes = express.Router(),
+    sipRoutes = express.Router(),
+    queueRoutes = express.Router(),
+    channelsRoutes = express.Router(),
+    userRoutes = express.Router(),
+    authRoutes = express.Router();
 
   //=========================
   // Auth Routes
@@ -65,7 +65,12 @@ module.exports = function (app) {
   // Set user routes as a subgroup/middleware to apiRoutes
   apiRoutes.use('/channels', channelsRoutes);
 
-  channelsRoutes.get('/', requireAuth, ChannelsController.getChannels, CommonController.getDefaultResponse);
+  channelsRoutes.get(
+    '/',
+    requireAuth,
+    ChannelsController.getChannels,
+    CommonController.getDefaultResponse
+  );
 
   channelsRoutes.get(
     '/spy/:recipient/:sip',
