@@ -8,24 +8,24 @@ import {
 
 export function getQueues() {
   return {
-    type    : LOAD_ALL_QUEUES,
-    callAPI : `${API_URL}/queue`,
+    type: LOAD_ALL_QUEUES,
+    callAPI: `${API_URL}/queue`,
     needAuth: true
   };
 }
 
 export function queueJoin({ entry, id }) {
   return {
-    type    : QUEUE_JOIN,
-    payload : entry,
+    type: QUEUE_JOIN,
+    payload: entry,
     needAuth: true
   };
 }
 
 export function queueLeave({ entry, id }) {
   return {
-    type    : QUEUE_LEAVE,
-    payload : entry,
+    type: QUEUE_LEAVE,
+    payload: entry,
     needAuth: true
   };
 }
