@@ -1,19 +1,31 @@
 import React, { Component } from 'react';
 
-import Header from 'Component/Header/index';
+import Header from 'components/Header/index';
+import { ToastContainer } from 'react-toastify';
 
+import 'react-toastify/dist/ReactToastify.css';
 import './content.css';
 
 class Content extends Component {
   render() {
     return (
-        <div>
-          <Header />
+      <div>
+        <ToastContainer
+          position="top-right"
+          autoClose={10000}
+          hideProgressBar={false}
+          newestOnTop
+          closeOnClick
+          rtl={false}
+          pauseOnVisibilityChange
+          draggable
+          pauseOnHover
+        />
 
-          <div>
-            {this.props.children}
-          </div>
-        </div>
+        <Header />
+
+        <div>{this.props.children}</div>
+      </div>
     );
   }
 }
