@@ -2,10 +2,7 @@ import React, { Component } from 'react';
 
 class Footer extends Component {
   render() {
-    return (
-      <footer className="footer">
-      </footer>
-    );
+    return <footer className="footer" />;
   }
 }
 
