@@ -1,60 +1,48 @@
-import React from "react";
-import { Card, CardBody, Col, Row } from "reactstrap";
-import SipPeer from "Component/SipPeer";
-import QueueEntry from "Component/QueueEntry";
-import SipToolbar from "Component/SipToolbar";
+import React from 'react';
+import { ButtonGroup, Card, CardBody, Row } from 'reactstrap';
+
+import QueueEntryContainer from 'components/QueueEntry/QueueEntryContainer';
+import QueueMembersContainer from 'components/Queue/QueueMembersContainer';
+
+import styles from './QueueComponent.module.css';
+import { getFilteredMembersLengthState } from '../../store/selectors/queue';
+import connect from 'react-redux/es/connect/connect';
+
+// const QueueDetails = ({ queueDetail }) => (
+//   <small>
+//     <ul>
+//       <li>Answered - {queueDetail.completed}</li>
+//       <li>Missed - {queueDetail.abandoned}</li>
+//     </ul>
+//   </small>
+// );
 
 const QueueComponent = props => {
-  const {
-    queueMembers,
-    queueEntries,
-    entity: queueName,
-    queueDetail,
-    selectSip,
-    selectedSip
-  } = props;
+  const { entity, isEmpty } = props;
 
-
-  if (!queueMembers || !queueMembers.size) {
+  if (isEmpty) {
     return null;
   }
 
-  const members = queueMembers.entrySeq().map(([key, sipPeer]) => {
-    const isSelected = selectedSip && selectedSip === sipPeer.sip;
-
-    return (
-      <Col key={sipPeer.sip} md="4" className="col-sip">
-        <SipPeer
-          sipPeer={sipPeer}
-          memberOfQueue={queueName}
-          selectSip={selectSip}
-          isSelected={!!isSelected}
-        />
-      </Col>
-    );
-  });
-
-  const entries = queueEntries.entrySeq().map(([key, sipPeer]) => {
-    return <QueueEntry key={sipPeer.uniqueid} sipPeer={sipPeer}  />;
-  });
-
   return (
-    <Card className={"queue"}>
-      <SipToolbar queueEntity={queueName} />
-
-      {/*<small>*/}
-        {/*<ul>*/}
-          {/*<li>Answered - {queueDetail.completed}</li>*/}
-          {/*<li>Missed - {queueDetail.abandoned}</li>*/}
-        {/*</ul>*/}
-      {/*</small>*/}
+    <Card className={'queue'}>
       <CardBody>
         <Row>
+          <ButtonGroup className={styles.title}>
+            <span className="badge badge-info">Queue - {entity}</span>
+          </ButtonGroup>
+        </Row>
+
+        <Row>
           <div className="col-md-9">
-            <Row>{members}</Row>
+            <Row>
+              <QueueMembersContainer queueId={entity} />
+            </Row>
           </div>
           <div className="col-md-3 left-divider">
-            <Row>{entries}</Row>
+            <Row>
+              <QueueEntryContainer queueId={entity} />
+            </Row>
           </div>
         </Row>
       </CardBody>
@@ -62,4 +50,16 @@ const QueueComponent = props => {
   );
 };
 
-export default QueueComponent;
+// export default QueueComponent;
+
+function mapStateToProps(state, { entity }) {
+  const getMembersLength = getFilteredMembersLengthState();
+
+  return {
+    isEmpty: !getMembersLength(state, entity)
+  };
+}
+export default connect(
+  mapStateToProps,
+  {}
+)(QueueComponent);
