@@ -1,7 +1,7 @@
-import React, { Component } from 'react';
+import React, { PureComponent } from 'react';
 
 import { connect } from 'react-redux';
-import { getUser, patchUser, postUser, redirectComplete } from 'AC/user';
+import { getUser, patchUser, postUser } from 'AC/user';
 import { NavLink as RRNavLink } from 'react-router-dom';
 import { Redirect } from 'react-router';
 
@@ -9,7 +9,6 @@ import PropTypes from 'prop-types';
 
 import {
   Container,
-  Row,
   Button,
   Form,
   FormGroup,
@@ -20,9 +19,28 @@ import {
   Breadcrumb,
   BreadcrumbItem
 } from 'reactstrap';
-import FontAwesome from 'react-fontawesome';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import {
+  getUserErrorFieldState,
+  getUserIsErrorMsgState,
+  getUserIsErrorState,
+  getUserIsSubmittedState,
+  getUserState
+} from '../../store/selectors/user';
 
-class User extends Component {
+import { toast } from 'react-toastify';
+import styles from './User.module.css';
+
+class User extends PureComponent {
+  static propTypes = {
+    user: PropTypes.object,
+    isSubmitSuccess: PropTypes.bool,
+    isNew: PropTypes.bool,
+    isError: PropTypes.bool,
+    errorMessage: PropTypes.string,
+    errorField: PropTypes.string
+  };
+
   state = {
     firstName: '',
     lastName: '',
@@ -30,35 +48,9 @@ class User extends Component {
     role: 'Client',
     sip: '',
     _id: '',
-    password: ''
+    password: '',
+    isSubmitSuccess: false
   };
-
-  componentWillReceiveProps(nextProps) {
-    const { getUser, user, match, isNew } = nextProps;
-    if (isNew) {
-      return;
-    }
-
-    const username = match.params.id;
-
-    if (!user) {
-      return getUser(username);
-    }
-
-    const { _id, isLoading } = user;
-    if (!_id && !isLoading) {
-      return getUser(username);
-    }
-
-    this.setState({
-      firstName: user.firstName,
-      lastName: user.lastName,
-      username: user.username,
-      role: user.role,
-      sip: user.sip,
-      _id: user._id
-    });
-  }
 
   componentDidMount() {
     const { match, getUser, user, isNew } = this.props;
@@ -68,7 +60,6 @@ class User extends Component {
 
     const username = match.params.id;
 
-    const {} = this.props;
     if (!user) {
       return getUser(username);
     }
@@ -79,51 +70,134 @@ class User extends Component {
     }
 
     this.setState({
-      firstName: user.firstName,
-      lastName: user.lastName,
-      username: user.username,
-      role: user.role,
-      sip: user.sip
+      ...user
     });
   }
+
+  // componentWillReceiveProps(nextProps) {
+  //   const { getUser, user, match, isNew } = nextProps;
+  //
+  //   if (nextProps.isSubmitSuccess === true && this.props.isSubmitSuccess === false) {
+  //     toast.success('User successful created');
+  //
+  //     this.setState({ isSubmitSuccess: true });
+  //   }
+  //
+  //   if (nextProps.isError === true) {
+  //     toast.error(nextProps.errorMessage);
+  //   }
+  //
+  //   if (isNew) {
+  //     return;
+  //   }
+  //
+  //   const username = match.params.id;
+  //
+  //   if (!user) {
+  //     return getUser(username);
+  //   }
+  //
+  //   const { _id, isLoading } = user;
+  //   if (!_id && !isLoading) {
+  //     return getUser(username);
+  //   }
+  //
+  //   this.setState({
+  //     ...user
+  //   });
+  // }
+
+  static getDerivedStateFromProps(nextProps, prevState) {
+    if (nextProps.user && nextProps.user.username !== prevState.username) {
+      return {
+        firstName: nextProps.user.firstName,
+        lastName: nextProps.user.lastName,
+        username: nextProps.user.username,
+        role: nextProps.user.role,
+        sip: nextProps.user.sip,
+        _id: nextProps.user._id
+      };
+    }
+
+    return null;
+  }
+
+  componentDidUpdate(prevProps, prevState, prevContext) {
+    const { isNew } = this.props;
+    if (this.props.isSubmitSuccess === true && prevProps.isSubmitSuccess === false) {
+      const message = isNew ? 'User successful created' : 'User successful updated';
+
+      toast.success(message);
+
+      this.setState({ isSubmitSuccess: true });
+    }
+
+    if (this.props.isError === true) {
+      toast.error(this.props.errorMessage || 'Error has occurred');
+    }
+  }
+
+  handleChangeInput = ev => {
+    this.setState({
+      [ev.target.name]: ev.target.value
+    });
+  };
 
   handleSubmitForm = ev => {
     ev.preventDefault();
     const { firstName, lastName, sip, role, username, password } = this.state;
     const { isNew, patchUser, postUser } = this.props;
 
-    if (isNew) {
-      return postUser({
-        username,
-        firstName,
-        lastName,
-        sip,
-        role,
-        password
-      });
-    }
-
-    patchUser(username, {
+    const user = {
+      username,
       firstName,
       lastName,
       sip,
       role,
       password
-    });
+    };
+
+    if (isNew) {
+      return postUser(user);
+    }
+
+    patchUser(user);
   };
 
-  render() {
-    const { user, isNew, isError, errorMessage, errorField, shouldRedirect, redirectComplete } = this.props;
-    const invalidFields = { [errorField]: { valid: false } };
+  _renderFormItem({ label, field, type = 'text', isInvalid = false, ...attributes }) {
+    const { errorMessage } = this.props;
 
-    if (shouldRedirect) {
-      redirectComplete();
-      return <Redirect to={shouldRedirect}/>;
+    return (
+      <FormGroup row>
+        <Label for={field} md={2} className="text-capitalize">
+          {label || field}
+        </Label>
+        <Col md={10}>
+          <Input
+            {...attributes}
+            type={type}
+            name={field}
+            id={field}
+            value={this.state[field]}
+            invalid={isInvalid}
+            onChange={this.handleChangeInput}
+          />
+          {isInvalid && errorMessage && <FormFeedback>{errorMessage}</FormFeedback>}
+        </Col>
+      </FormGroup>
+    );
+  }
+
+  render() {
+    const { user, isNew, isError, errorMessage, errorField } = this.props;
+
+    if (this.state.isSubmitSuccess) {
+      return <Redirect to={'/users'} />;
     }
 
     return (
       <Container>
-        <Breadcrumb style={{ margin: '20px' }}>
+        <Breadcrumb className={styles['bread-crumbs']}>
           <BreadcrumbItem>
             <RRNavLink to="/users/">Users</RRNavLink>
           </BreadcrumbItem>
@@ -133,158 +207,137 @@ class User extends Component {
           </BreadcrumbItem>
         </Breadcrumb>
 
-        <div style={{ margin: '20px' }}>
+        <div className={styles.content}>
           <Form onSubmit={this.handleSubmitForm}>
             {isNew && (
-              <FormGroup row>
-                <Label for="username" md={2}>
-                  Username
-                </Label>
-                <Col md={10}>
-                  <Input
-                    disabled={!isNew}
-                    type="text"
-                    name="username"
-                    id="username"
-                    value={this.state.username}
-                    {...invalidFields['username']}
-                    // required={true}
-                    onChange={ev => this.setState({ username: ev.target.value })}
-                  />
-                  {isError &&
-                  errorField === 'username' && <FormFeedback>{errorMessage}</FormFeedback>}
-                </Col>
-              </FormGroup>
+              <FormItem
+                value={this.state.username}
+                field="username"
+                isInvalid={isError && errorField === 'username'}
+                errorMessage={errorMessage}
+                handleOnChange={this.handleChangeInput}
+              />
             )}
 
-            <FormGroup row>
-              <Label for="password" md={2}>
-                Password
-              </Label>
-              <Col md={10}>
-                <Input
-                  type="password"
-                  name="password"
-                  id="password"
-                  autoComplete="new-password"
-                  value={this.state.password}
-                  {...invalidFields['password']}
-                  // required={true}
-                  onChange={ev => this.setState({ password: ev.target.value })}
-                />
-                {isError &&
-                errorField === 'password' && <FormFeedback>{errorMessage}</FormFeedback>}
-              </Col>
-            </FormGroup>
+            <FormItem
+              type="password"
+              value={this.state.password}
+              field="password"
+              isInvalid={isError && errorField === 'password'}
+              errorMessage={errorMessage}
+              handleOnChange={this.handleChangeInput}
+            />
 
-            <FormGroup row>
-              <Label for="role" md={2}>
-                Role
-              </Label>
-              <Col md={10}>
-                <Input
-                  type="select"
-                  name="role"
-                  id="role"
-                  value={this.state.role}
-                  {...invalidFields['role']}
-                  onChange={ev => this.setState({ role: ev.target.value })}
-                >
-                  <option value={'Admin'}>Admin</option>
-                  <option value={'Client'}>Client</option>
-                </Input>
-                {isError && errorField === 'role' && <FormFeedback>{errorMessage}</FormFeedback>}{' '}
-              </Col>
-            </FormGroup>
+            <FormItem
+              value={this.state.role}
+              field="role"
+              type="select"
+              isInvalid={isError && errorField === 'role'}
+              errorMessage={errorMessage}
+              handleOnChange={this.handleChangeInput}
+              options={[{ key: 'Admin' }, { key: 'Client' }]}
+            />
 
-            <FormGroup row>
-              <Label for="firstName" md={2}>
-                First Name
-              </Label>
-              <Col md={10}>
-                <Input
-                  type="text"
-                  name="firstName"
-                  id="firstName"
-                  value={this.state.firstName}
-                  {...invalidFields['firstName']}
-                  // required={true}
-                  onChange={ev => this.setState({ firstName: ev.target.value })}
-                />
-                {isError && errorField === 'firstName' && <FormFeedback>{errorMessage}</FormFeedback>}{' '}
-              </Col>
-            </FormGroup>
+            <FormItem
+              value={this.state.firstName}
+              field="firstName"
+              label="First Name"
+              isInvalid={isError && errorField === 'firstName'}
+              errorMessage={errorMessage}
+              handleOnChange={this.handleChangeInput}
+            />
 
-            <FormGroup row>
-              <Label for="lastName" md={2}>
-                Last Name
-              </Label>
-              <Col md={10}>
-                <Input
-                  type="text"
-                  name="lastName"
-                  id="lastName"
-                  value={this.state.lastName}
-                  {...invalidFields['lastName']}
-                  // required={true}
-                  onChange={ev => this.setState({ lastName: ev.target.value })}
-                />
-                {isError && errorField === 'lastName' && <FormFeedback>{errorMessage}</FormFeedback>}{' '}
-              </Col>
-            </FormGroup>
+            <FormItem
+              value={this.state.lastName}
+              field="lastName"
+              label="Last Name"
+              isInvalid={isError && errorField === 'lastName'}
+              errorMessage={errorMessage}
+              handleOnChange={this.handleChangeInput}
+            />
 
-            <FormGroup row>
-              <Label for="sip" md={2}>
-                SIP
-              </Label>
-              <Col md={10}>
-                <Input
-                  type="text"
-                  name="sip"
-                  id="sip"
-                  value={this.state.sip}
-                  {...invalidFields['sip']}
-                  // required={true}
-                  onChange={ev => this.setState({ sip: ev.target.value })}
-                />
-                {isError && errorField === 'sip' && <FormFeedback>{errorMessage}</FormFeedback>}{' '}
-              </Col>
-            </FormGroup>
+            <FormItem
+              value={this.state.sip}
+              field="sip"
+              isInvalid={isError && errorField === 'sip'}
+              errorMessage={errorMessage}
+              handleOnChange={this.handleChangeInput}
+            />
 
             {isNew && <Button>Create User</Button>}
 
             {!isNew && (
               <Button disabled={!user}>
                 Change &nbsp;
-                {(!user || user.isLoading) && <FontAwesome name="circle-o-notch" spin />}
+                {(!user || user.isLoading) && <FontAwesomeIcon icon="circle-notch" spin />}
               </Button>
             )}
           </Form>
         </div>
       </Container>
     );
-
   }
 }
 
-User.propTypes = {
-  user: PropTypes.object
+const FormItem = ({
+  value,
+  label,
+  field,
+  type = 'text',
+  isInvalid = false,
+  errorMessage,
+  handleOnChange,
+  ...attributes
+}) => {
+  const { options, ...rest } = attributes;
+  let Options = null;
+
+  if (options) {
+    Options = options.map(el => (
+      <option key={el.key} value={el.key}>
+        {el.label || el.key}
+      </option>
+    ));
+  }
+
+  return (
+    <FormGroup row>
+      <Label for={field} md={2} className="text-capitalize">
+        {label || field}
+      </Label>
+      <Col md={10}>
+        <Input
+          {...rest}
+          type={type}
+          name={field}
+          id={field}
+          value={value}
+          invalid={isInvalid}
+          onChange={handleOnChange}
+        >
+          {Options}
+        </Input>
+        {isInvalid && errorMessage && <FormFeedback>{errorMessage}</FormFeedback>}
+      </Col>
+    </FormGroup>
+  );
 };
 
 function mapStateToProps(state, { match }) {
   const username = match.params.id;
-
   const isNew = username === 'register';
-  const user = !isNew ? state.user.getIn(['entities', username]) : null;
 
   return {
-    user,
-    isNew,
-    isError: state.user.get('isError'),
-    errorMessage: state.user.get('errorMessage'),
-    errorField: state.user.get('errorField'),
-    shouldRedirect: state.user.get('shouldRedirect'),
+    user: !isNew ? getUserState(state, username) : null,
+    isNew: isNew,
+    isError: getUserIsErrorState(state),
+    errorMessage: getUserIsErrorMsgState(state),
+    errorField: getUserErrorFieldState(state),
+    isSubmitSuccess: getUserIsSubmittedState(state)
   };
 }
 
-export default connect(mapStateToProps, { getUser, patchUser, postUser, redirectComplete })(User);
+export default connect(
+  mapStateToProps,
+  { getUser, patchUser, postUser }
+)(User);
