@@ -1,4 +1,5 @@
-import { combineReducers } from 'redux';
+import { combineReducers } from 'redux-immutable';
+
 import auth from './auth';
 import sip from './sip';
 import queue from './queue';
